Move runNonDetOnce back into the NonDet module.
