+++ src/Control/Effect.hs
@@ -10,7 +10,7 @@ import Control.Effect.Fail      as X (Fail, FailC)
 import Control.Effect.Fresh     as X (Fresh, FreshC)
 import Control.Effect.Internal  as X (Eff, interpret)
 import Control.Effect.Lift      as X (Lift, LiftC, runM)
-import Control.Effect.NonDet    as X (NonDet, AltC)
+import Control.Effect.NonDet    as X (NonDet, AltC, OnceC)
 import Control.Effect.Random    as X (Random, RandomC)
 import Control.Effect.Reader    as X (Reader, ReaderC)
 import Control.Effect.Resource  as X (Resource, ResourceC)
